Only use course ID in course directory name
Including the course name in the local course directory name is a
bad idea because course names can contain slashes. As a result, we
revert to the previous naming schema where a course directory's name
is the ID of the corresponding course.

We now store the course name itself in a file called "course_name.txt"
within the course directory.

No action is required from existing bot users.